move graph/network to deprecated
