+++ src/python/dxl/learn/core/__init__.py
@@ -2,6 +2,5 @@ from .config import ConfigurableWithName, ConfigurableWithClass, set_global_conf
 from .graph_info import GraphInfo
 from .session import make_session, ThisSession, Session, SessionBase
 from .tensor import Tensor, Variable, Constant, NoOp, NotTrainableVariable
-from .graph import Graph
 from .model import Model
 from .subgraph_maker import SubgraphMakerFactory, SubgraphMakerFinder, SubgraphMakerTable, SubgraphPartialMaker
