+++ src/python/dxl/learn/core/model.py
@@ -1,8 +1,6 @@
-from .graph import Graph
-from .tensor import Tensor, Constant
-from .graph_info import GraphInfo
-from typing import Dict, Union, Tuple
-from dxl.fs import Path
+from src.python.dxl.learn.deprecated.graph import Graph
+from .tensor import Tensor
+from typing import Dict
 import tensorflow as tf
