--- src/python/dxl/learn/network/__init__.py
@@ -1,1 +1,0 @@
-from .network import Network
